+++ tests/BreachTest.php
@@ -10,6 +10,7 @@ namespace Icawebdesign\Test;
 
 use Icawebdesign\Hibp\Breach\Breach;
 use Icawebdesign\Hibp\Breach\BreachSiteEntity;
+use Icawebdesign\Hibp\Breach\BreachSiteTruncatedEntity;
 use PHPUnit\Framework\TestCase;
 
 class BreachTest extends TestCase
@@ -27,11 +28,12 @@ class BreachTest extends TestCase
 
     public function tearDown(): void
     {
+        parent::tearDown();
         $this->breach = null;
     }
 
     /** @test */
-    public function gettingAllBreachesitesReturnsACollection()
+    public function gettingAllBreachSitesReturnsACollection(): void
     {
         $breaches = $this->breach->getAllBreachSites();
 
@@ -42,7 +44,18 @@ class BreachTest extends TestCase
     }
 
     /** @test */
-    public function successfulBreachLookupReturnsABreachSiteEntity()
+    public function gettingAllFilteredBreachSitesReturnsACollection(): void
+    {
+        $breaches = $this->breach->getAllBreachSites('adobe.com');
+
+        $this->assertEquals(200, $this->breach->getStatusCode());
+        $this->assertInstanceOf(\Tightenco\Collect\Support\Collection::class, $breaches);
+        $this->assertGreaterThan(0, $breaches->count());
+        $this->assertInstanceOf(BreachSiteEntity::class, $breaches->first());
+    }
+
+    /** @test */
+    public function successfulBreachLookupReturnsABreachSiteEntity(): void
     {
         $breachedAccount = $this->breach->getBreach('000webhost');
 
@@ -65,7 +78,7 @@ class BreachTest extends TestCase
     }
 
     /** @test */
-    public function unsuccessfulBreachLookupThrowsABreachNotFoundException()
+    public function unsuccessfulBreachLookupThrowsABreachNotFoundException(): void
     {
         $this->expectException(\Icawebdesign\Hibp\Exception\BreachNotFoundException::class);
 
@@ -73,7 +86,7 @@ class BreachTest extends TestCase
     }
 
     /** @test */
-    public function gettingAllDataclassesReturnsACollection()
+    public function gettingAllDataclassesReturnsACollection(): void
     {
         $dataClasses = $this->breach->getAllDataClasses();
         $this->assertEquals(200, $this->breach->getStatusCode());
@@ -81,9 +94,9 @@ class BreachTest extends TestCase
     }
 
     /** @test */
-    public function gettingBreachDataForAccountReturnsACollection()
+    public function gettingBreachDataForAccountReturnsACollection(): void
     {
-        $breaches = $this->breach->getBreachedAccount('test@example.com');
+        $breaches = $this->breach->getBreachedAccount('test@example.com', false);
 
         $this->assertEquals(200, $this->breach->getStatusCode());
         $this->assertInstanceOf(\Tightenco\Collect\Support\Collection::class, $breaches);
@@ -92,9 +105,50 @@ class BreachTest extends TestCase
     }
 
     /** @test */
-    public function gettingBreachDataForAnInvalidAccountThrowsABreachNotFoundException()
+    public function gettingBreachDataForAnInvalidAccountThrowsABreachNotFoundException(): void
     {
         $this->expectException(\Icawebdesign\Hibp\Exception\BreachNotFoundException::class);
         $this->breach->getBreachedAccount('invalid_email_address');
+    }
+
+    /** @test */
+    public function gettingTruncatedBreachedAccountsReturnsACollectionOfBreachSiteTruncatedEntities(): void
+    {
+        $breaches = $this->breach->getBreachedAccountTruncated('test@example.com', false);
+
+        $this->assertEquals(200, $this->breach->getStatusCode());
+        $this->assertInstanceOf(\Tightenco\Collect\Support\Collection::class, $breaches);
+        $this->assertGreaterThan(0, $breaches->count());
+        $this->assertInstanceOf(BreachSiteTruncatedEntity::class, $breaches->first());
+    }
+
+    /** @test */
+    public function gettingFilteredBreachedAccountReturnsACollectionOfBreachSiteEntities(): void
+    {
+        $breaches = $this->breach->getBreachedAccount(
+            'test@example.com',
+            true,
+            'adobe.com'
+        );
+
+        $this->assertEquals(200, $this->breach->getStatusCode());
+        $this->assertInstanceOf(\Tightenco\Collect\Support\Collection::class, $breaches);
+        $this->assertGreaterThan(0, $breaches->count());
+        $this->assertInstanceOf(BreachSiteEntity::class, $breaches->first());
+    }
+
+    /** @test */
+    public function gettingFilteredTruncatedBreachedAccountsReturnsACollectionOfBreachSiteTruncatedEntities(): void
+    {
+        $breaches = $this->breach->getBreachedAccountTruncated(
+            'test@example.com',
+            false,
+            'adobe.com'
+        );
+
+        $this->assertEquals(200, $this->breach->getStatusCode());
+        $this->assertInstanceOf(\Tightenco\Collect\Support\Collection::class, $breaches);
+        $this->assertGreaterThan(0, $breaches->count());
+        $this->assertInstanceOf(BreachSiteTruncatedEntity::class, $breaches->first());
     }
 }