+++ tests/PwnedPasswordTest.php
@@ -31,7 +31,8 @@ class PwnedPasswordTest extends TestCase
     /** @test */
     public function successfulRangeLookupReturnsAPositiveInteger()
     {
-        $response = $this->pwnedPassword->range('5baa6', '5BAA61E4C9B93F3F0682250B6CF8331B7EE68FD8');
+//        $response = $this->pwnedPassword->range('5baa6', '5BAA61E4C9B93F3F0682250B6CF8331B7EE68FD8');
+        $response = $this->pwnedPassword->rangeFromHash('5BAA61E4C9B93F3F0682250B6CF8331B7EE68FD8');
 
         $this->assertEquals(200, $this->pwnedPassword->getStatusCode());
         $this->assertIsInt($response);
@@ -41,7 +42,7 @@ class PwnedPasswordTest extends TestCase
     /** @test */
     public function failedRangeLookupReturnsZero()
     {
-        $response = $this->pwnedPassword->range('00000', '0000000000000000000000000000000000000000');
+        $response = $this->pwnedPassword->rangeFromHash('0000000000000000000000000000000000000000');
 
         $this->assertEquals(200, $this->pwnedPassword->getStatusCode());
         $this->assertEquals(0, $response);
@@ -50,7 +51,7 @@ class PwnedPasswordTest extends TestCase
     /** @test */
     public function successfulRangeDataLookupReturnsACollection()
     {
-        $response = $this->pwnedPassword->rangeData('5baa6', '5baa61e4c9b93f3f0682250b6cf8331b7ee68fd8');
+        $response = $this->pwnedPassword->rangeDataFromHash('5baa61e4c9b93f3f0682250b6cf8331b7ee68fd8');
 
         $this->assertEquals(200, $this->pwnedPassword->getStatusCode());
         $this->assertInstanceOf(\Tightenco\Collect\Support\Collection::class, $response);
@@ -61,7 +62,7 @@ class PwnedPasswordTest extends TestCase
     {
         $this->expectException(\GuzzleHttp\Exception\RequestException::class);
 
-        $this->pwnedPassword->range('&&', '&&');
+        $this->pwnedPassword->rangeFromHash('&&');
     }
 
     /** @test */
@@ -69,6 +70,6 @@ class PwnedPasswordTest extends TestCase
     {
         $this->expectException(\GuzzleHttp\Exception\RequestException::class);
 
-        $this->pwnedPassword->rangeData('&&', '&&');
+        $this->pwnedPassword->rangeDataFromHash('&&');
     }
 }